Use bufio.Scanner instead of channel for reading client data
A Scanner is much better suited to reading stream data. This also eliminates
an unnecessary goroutine whose sole purpose was to scan lines and push
them into a channel. Additionally, this opens the path for certain request handlers
to read directly from the connection io.Reader where necessary. This will be needed
for making use of the textproto.Reader.ReadMIMEHeader method