Fix tests, change Serve and remove Close
A couple of broken tests have been fixed. This was due to refactored
changes not being reflected everywhere. Serve was changed so that
it's closer to that in net/http (as I think was the original
intention). It takes a listener. This makes it possible to listen on
multiple interfaces and ports using the same server. Close was removed
as it's unnecessary now.